Fix issues storing mp_id when using MariaDB
Since IntegerField has a limit of 2^31 on MariaDB, that doesn't work for
saving the external ids for Payments. Use BigInteger instead.

Fixes #8

+++ django_mercadopago/models.py
@@ -346,7 +346,7 @@ class Payment(models.Model):
     A payment received, related to a preference.
     """
 
-    mp_id = models.IntegerField(
+    mp_id = models.BigIntegerField(
         _('mp id'),
         unique=True,
     )
